requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
django-debug-toolbar>=4.0
